AMQPClient has been removed. EventHubsClientWrapper is the only client implementation
